+++ test/integration/node-specific/errors.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { MongoClient, MongoError, MongoServerSelectionError } from '../../mongodb';
+import { MongoClient, MongoError, MongoServerSelectionError, ReadPreference } from '../../mongodb';
 
 describe('Error (Integration)', function () {
   describe('AggregateErrors', function () {
@@ -50,5 +50,38 @@ describe('Error (Integration)', function () {
     }).catch(e => e);
     expect(error).to.be.instanceOf(MongoServerSelectionError);
     expect(error.message).not.to.be.empty;
+  });
+
+  context('when a server selection error is stringified', function () {
+    it(
+      'the error"s topology description correctly displays the `servers`',
+      { requires: { topology: 'replicaset' } },
+      async function () {
+        const client: MongoClient = this.configuration.newClient({
+          serverSelectionTimeoutMS: 1000
+        });
+        try {
+          await client.connect();
+
+          const error = await client
+            .db('foo')
+            .collection('bar')
+            .find(
+              {},
+              {
+                // Use meaningless read preference tags to ensure that the server selection fails
+                readPreference: new ReadPreference('secondary', [{ ny: 'ny' }])
+              }
+            )
+            .toArray()
+            .catch(e => JSON.parse(JSON.stringify(e)));
+
+          const servers = error.reason.servers;
+          expect(Object.keys(servers).length > 0).to.be.true;
+        } finally {
+          await client.close();
+        }
+      }
+    );
   });
 });
